+++ nidup/pysc2/agent/scripted/commander.py
@@ -1,6 +1,6 @@
 
 from nidup.pysc2.wrapper.observations import Observations
-from nidup.pysc2.order import Order
+from nidup.pysc2.agent.order import Order
 from nidup.pysc2.information import BaseLocation, StepIndex
 from nidup.pysc2.agent.scripted.build import OrdersSequence, CenterCameraOnCommandCenter, BuildSupplyDepot, BuildFactory, BuildRefinery, BuildBarracks, BuildTechLabBarracks, MorphOrbitalCommand
 from nidup.pysc2.agent.scripted.train import OrdersRepetition, TrainMarine, TrainMarauder, PushWithArmy
