+++ nidup/pysc2/agent/scripted/scout.py
@@ -2,7 +2,7 @@
 from pysc2.lib import actions
 from nidup.pysc2.wrapper.actions import TerranActions, TerranActionIds
 from nidup.pysc2.wrapper.observations import Observations
-from nidup.pysc2.order import Order
+from nidup.pysc2.agent.order import Order
 from nidup.pysc2.information import BaseLocation
 from nidup.pysc2.wrapper.unit_types import UnitTypeIds
 import random
